requirements.txt
asn1crypto==1.5.1
boto3==1.42.16
botocore==1.42.16
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dlt==1.20.0
filelock==3.20.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.14.0
humanize==4.15.0
idna==3.11
jmespath==1.0.1
jsonpath-ng==1.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathvalidate==3.3.1
pendulum==3.1.0
platformdirs==4.5.1
pluggy==1.6.0
ply==3.11
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requirements-parser==0.13.0
rich==14.2.0
rich-argparse==1.7.2
s3transfer==0.16.0
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.1.1
sortedcontainers==2.4.0
sqlglot==28.5.0
tenacity==9.1.2
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
